Updating for POST method, version with HTTP/1.1
- Updated POST method working
- Updated HTTP version compatibility for 1.1

+++ src/shttptool.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 const char* SimpleHTTPTool::GetMIME( SHTTPContentType ctype )
 {
-    unsigned int   mjr = (unsigned int)ctype & 0xFFFF0000;
+    unsigned long  mjr = (unsigned int)ctype & 0xFFFF0000;
     unsigned short mnr = (unsigned int)ctype & 0x0000FFFF;
 
     return GetMIME( (SHTTPCT_MAJOR)mjr, (SHTTPCT_MINOR)mnr );
